requirements.txt
pydantic>=2.8
pyyaml>=6.0.1
jinja2>=3.1.4
rich>=13.7.1
pytest>=8.3.2
